git-svn-id: https://funk2.googlecode.com/svn/funk2@665 98e8950a-0dbf-11de-ba4d-252e5d5a8945

+++ c/f2_global.h
@@ -70,7 +70,7 @@ typedef                          u8 pool_index_t;
 #  define pool_index__fstr       X8__fstr
 #  define pool_index__bit_num     1
 
-typedef                         u64 pool_address_t;
+typedef                         u32 pool_address_t;
 #  define pool_address__fstr    X64__fstr
 #  define pool_address__bit_num  30
